inference-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiohttp>=3.8.0
msal>=1.22.0
google-cloud-aiplatform>=1.30.0
google-cloud-discoveryengine>=0.11.0
python-multipart>=0.0.6
jinja2>=3.1.2
requests>=2.31.0
streamlit>=1.32.0
